+++ physicsnemo/utils/sdf.py
@@ -14,8 +14,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import cupy as cp
-import numpy as np
+import torch
 import warp as wp
 
 wp.config.quiet = True
@@ -28,7 +27,6 @@ def _bvh_query_distance(
     max_dist: wp.float32,
     sdf: wp.array(dtype=wp.float32),
     sdf_hit_point: wp.array(dtype=wp.vec3f),
-    sdf_hit_point_id: wp.array(dtype=wp.int32),
     use_sign_winding_number: bool = False,
 ):
     """
@@ -67,22 +65,16 @@ def _bvh_query_distance(
 
     sdf[tid] = res.sign * wp.abs(wp.length(points[tid] - p_closest))
     sdf_hit_point[tid] = p_closest
-    sdf_hit_point_id[tid] = res.face
 
 
-Array = np.ndarray | cp.ndarray
-
-
+@torch.library.custom_op("physicsnemo::signed_distance_field", mutates_args=())
 def signed_distance_field(
-    mesh_vertices: Array,
-    mesh_indices: Array,
-    input_points: Array,
+    mesh_vertices: torch.Tensor,
+    mesh_indices: torch.Tensor,
+    input_points: torch.Tensor,
     max_dist: float = 1e8,
-    include_hit_points: bool = False,
-    include_hit_points_id: bool = False,
     use_sign_winding_number: bool = False,
-    return_cupy: bool | None = None,
-) -> Array | tuple[Array, ...]:
+) -> tuple[torch.Tensor, torch.Tensor]:
     """
     Computes the signed distance field (SDF) for a given mesh and input points.
 
@@ -100,11 +92,7 @@ def signed_distance_field(
         max_dist (float, optional): Maximum distance within which
             to search for the closest point on the mesh. Default is 1e8.
         include_hit_points (bool, optional): Whether to include hit points in
-            the output. Here, "hit points" are the points on the mesh that are
-            closest to the input points, and hence, are defining the SDF.
-            Default is False.
-        include_hit_points_id (bool, optional): Whether to include hit point
-            IDs in the output. Default is False.
+            the output. Here,
         use_sign_winding_number (bool, optional): Whether to use sign winding
             number method for SDF. Default is False. If False, your mesh should
             be watertight to obtain correct results.
@@ -115,88 +103,103 @@ def signed_distance_field(
     Returns:
     -------
     Returns:
-        np.ndarray | cp.ndarray or tuple:
-            - If both `include_hit_points` and `include_hit_points_id` are False
-              (default), returns a 1D array of signed distances for each input
-              point.
-            - If `include_hit_points` is True, returns a tuple: (sdf,
-              hit_points), where `hit_points` contains the closest mesh point
-              for each input point.
-            - If `include_hit_points_id` is True, returns a tuple: (sdf,
-              hit_point_ids), where `hit_point_ids` contains the face index of
-              the closest mesh face for each input point.
-            - If both `include_hit_points` and `include_hit_points_id` are True,
-              returns a tuple: (sdf, hit_points, hit_point_ids).
-            - The returned array type (NumPy or CuPy) is determined by the
-            `return_cupy` argument, or inferred from the input arrays.
+        tuple[torch.Tensor, torch.Tensor] of:
+            - signed distance to the mesh, per input point
+            - hit point, per input point. "hit points" are the points on the
+              mesh that are closest to the input points, and hence, are
+              defining the SDF.
 
     Example:
     -------
     >>> mesh_vertices = [(0, 0, 0), (1, 0, 0), (0, 1, 0)]
-    >>> mesh_indices = np.array((0, 1, 2))
-    >>> input_points = [(0.5, 0.5, 0.5)]
+    >>> mesh_indices = torch.tensor((0, 1, 2))
+    >>> input_points = torch.tensor((0.5, 0.5, 0.5))
     >>> signed_distance_field(mesh_vertices, mesh_indices, input_points)
-    array([0.5], dtype=float32)
+    (tensor([0.5]), tensor([0.5, 0.5, 0.5]))
     """
-    if return_cupy is None:
-        return_cupy = any(
-            isinstance(arr, cp.ndarray)
-            for arr in (mesh_vertices, mesh_indices, input_points)
+
+    if input_points.shape[-1] != 3:
+        raise ValueError("Input points must be a tensor with last dimension of size 3")
+
+    input_shape = input_points.shape
+
+    # Flatten the input points:
+    input_points = input_points.reshape(-1, 3)
+
+    N = len(input_points)
+
+    # Allocate output tensors with torch:
+    sdf = torch.zeros(N, dtype=torch.float32, device=input_points.device)
+    sdf_hit_point = torch.zeros(N, 3, dtype=torch.float32, device=input_points.device)
+
+    if input_points.device.type == "cuda":
+        wp_launch_stream = wp.stream_from_torch(
+            torch.cuda.current_stream(input_points.device)
         )
-
-    wp.init()
-
-    if isinstance(mesh_vertices, cp.ndarray):
-        device = mesh_vertices.device
-        wp_device = f"cuda:{device.id}"
+        wp_launch_device = None  # We explicitly pass None if using the stream.
     else:
-        wp_device = wp.get_device()
+        wp_launch_stream = None
+        wp_launch_device = "cpu"  # CPUs have no streams
 
-    with wp.ScopedDevice(wp_device):
+    with wp.ScopedStream(wp_launch_stream):
+        wp.init()
+
+        # zero copy the vertices, indices, and input points to warp:
+        wp_vertices = wp.from_torch(mesh_vertices.to(torch.float32), dtype=wp.vec3)
+        wp_indices = wp.from_torch(mesh_indices.to(torch.int32), dtype=wp.int32)
+        wp_input_points = wp.from_torch(input_points.to(torch.float32), dtype=wp.vec3)
+
+        # Convert output points:
+        wp_sdf = wp.from_torch(sdf, dtype=wp.float32)
+        wp_sdf_hit_point = wp.from_torch(sdf_hit_point, dtype=wp.vec3f)
+
         mesh = wp.Mesh(
-            points=wp.array(mesh_vertices, dtype=wp.vec3f, device=wp_device),
-            indices=wp.array(mesh_indices, dtype=wp.int32, device=wp_device),
+            points=wp_vertices,
+            indices=wp_indices,
+            support_winding_number=use_sign_winding_number,
         )
-
-        warp_input_points = wp.array(input_points, dtype=wp.vec3f, device=wp_device)
-
-        N = len(warp_input_points)
-
-        sdf = wp.empty(shape=(N,), dtype=wp.float32, device=wp_device)
-        sdf_hit_point = wp.empty(shape=(N,), dtype=wp.vec3f, device=wp_device)
-        sdf_hit_point_id = wp.empty(shape=(N,), dtype=wp.int32, device=wp_device)
 
         wp.launch(
             kernel=_bvh_query_distance,
             dim=N,
             inputs=[
                 mesh.id,
-                warp_input_points,
+                wp_input_points,
                 max_dist,
-                sdf,
-                sdf_hit_point,
-                sdf_hit_point_id,
+                wp_sdf,
+                wp_sdf_hit_point,
                 use_sign_winding_number,
             ],
-            device=wp_device,
+            device=wp_launch_device,
+            stream=wp_launch_stream,
         )
 
-        def convert(array: wp.array) -> np.ndarray | cp.ndarray:
-            """Converts a Warp array to CuPy/NumPy based on the `return_cupy` flag."""
-            if return_cupy:
-                return cp.asarray(array)
-            else:
-                return array.numpy()
+    # Unflatten the output to be like the input:
+    sdf = sdf.reshape(input_shape[:-1])
+    sdf_hit_point = sdf_hit_point.reshape(input_shape)
 
-        arrays_to_return: list[np.ndarray | cp.ndarray] = [convert(sdf)]
+    return sdf.to(input_points.dtype), sdf_hit_point.to(input_points.dtype)
 
-        if include_hit_points:
-            arrays_to_return.append(convert(sdf_hit_point))
-        if include_hit_points_id:
-            arrays_to_return.append(convert(sdf_hit_point_id))
 
-        return (
-            arrays_to_return[0]
-            if len(arrays_to_return) == 1
-            else tuple(arrays_to_return)
-        )
+@signed_distance_field.register_fake
+def _(
+    mesh_vertices: torch.Tensor,
+    mesh_indices: torch.Tensor,
+    input_points: torch.Tensor,
+    max_dist: float = 1e8,
+    use_sign_winding_number: bool = False,
+) -> tuple[torch.Tensor, torch.Tensor]:
+    if mesh_vertices.device != input_points.device:
+        raise RuntimeError("mesh_vertices and input_points must be on the same device")
+
+    if mesh_vertices.device != mesh_indices.device:
+        raise RuntimeError("mesh_vertices and mesh_indices must be on the same device")
+
+    N = input_points.shape[0]
+
+    sdf_output = torch.empty(N, 1, device=input_points.device, dtype=input_points.dtype)
+    sdf_hit_point_output = torch.empty(
+        N, 3, device=input_points.device, dtype=input_points.dtype
+    )
+
+    return sdf_output, sdf_hit_point_output